requirements.txt
fastapi~=0.115.4
uvicorn
pandas~=2.2.3
numpy~=2.1.2
scikit-learn~=1.5.2
requests~=2.32.3
python-dotenv~=1.0.1
app~=0.0.1
pydantic~=2.9.2